replaced old runnc with new runnc-cli
