--- runnc-cli/README.md
@@ -1,3 +1,0 @@
-This folder is to contain runnc cli implementation which will not use runc.
-
-The structure of the code is based around that of the original [`runc`](https://github.com/opencontainers/runc).
