--- runnc-cli/local.log
@@ -1,1 +1,0 @@
-time="2018-08-20T13:31:30-04:00" level=error msg="JSON specification file config.json not found"
