--- runnc-cli/runnc.go
@@ -1,155 +1,0 @@
-// Copyright 2014 Docker, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package main
-
-import (
-	"fmt"
-	"io"
-	"os"
-	"strings"
-
-	"github.com/opencontainers/runtime-spec/specs-go"
-
-	"github.com/sirupsen/logrus"
-	"github.com/urfave/cli"
-)
-
-// version will be populated by the Makefile, read from
-// VERSION file of the source code.
-var version = ""
-
-// gitCommit will be the hash that the binary was built from
-// and will be populated by the Makefile
-var gitCommit = ""
-
-const (
-	specConfig = "config.json"
-	usage      = `Nabla Containers runtime
-
-runnc is a command line client for running applications packaged according to
-the Open Container Initiative (OCI) format.
-
-Containers are configured using bundles. A bundle for a container is a directory
-that includes a specification file named "` + specConfig + `" and a root filesystem.
-The root filesystem contains the contents of the container.
-
-To start a new instance of a container:
-
-    # runnc run [ -b bundle ] <container-id>
-
-Where "<container-id>" is your name for the instance of the container that you
-are starting. The name you provide for the container instance must be unique on
-your host. Providing the bundle directory using "-b" is optional. The default
-value for "bundle" is the current directory.`
-)
-
-func main() {
-	app := cli.NewApp()
-	app.Name = "runnc"
-	app.Usage = usage
-
-	var v []string
-	if version != "" {
-		v = append(v, version)
-	}
-	if gitCommit != "" {
-		v = append(v, fmt.Sprintf("commit: %s", gitCommit))
-	}
-	v = append(v, fmt.Sprintf("spec: %s", specs.Version))
-	app.Version = strings.Join(v, "\n")
-
-	root := "/run/runnc"
-
-	app.Flags = []cli.Flag{
-		cli.BoolFlag{
-			Name:  "debug",
-			Usage: "enable debug output for logging",
-		},
-		cli.StringFlag{
-			Name:  "log",
-			Value: "/dev/null",
-			Usage: "set the log file path where internal debug information is written",
-		},
-		cli.StringFlag{
-			Name:  "log-format",
-			Value: "text",
-			Usage: "set the format used by logs ('text' (default), or 'json')",
-		},
-		cli.StringFlag{
-			Name:  "root",
-			Value: root,
-			Usage: "root directory for storage of container state (this should be located in tmpfs)",
-		},
-	}
-	app.Commands = []cli.Command{
-		// Implement essentials first (for baseic docker run to work)
-		createCommand,
-		deleteCommand,
-		stateCommand,
-		startCommand,
-		killCommand,
-		//		checkpointCommand,
-		//		eventsCommand,
-		//		execCommand,
-		initCommand,
-		//		listCommand,
-		//		pauseCommand,
-		//		psCommand,
-		//		restoreCommand,
-		//		resumeCommand,
-		//		runCommand,
-		//		specCommand,
-		//		updateCommand,
-	}
-	app.Before = func(context *cli.Context) error {
-		if context.GlobalBool("debug") {
-			logrus.SetLevel(logrus.DebugLevel)
-		}
-		if path := context.GlobalString("log"); path != "" {
-			f, err := os.OpenFile(path, os.O_CREATE|os.O_WRONLY|os.O_APPEND|os.O_SYNC, 0666)
-			if err != nil {
-				fmt.Fprintln(os.Stdout, err.Error())
-				return err
-			}
-			logrus.SetOutput(f)
-		}
-		switch context.GlobalString("log-format") {
-		case "text":
-			// retain logrus's default.
-		case "json":
-			logrus.SetFormatter(new(logrus.JSONFormatter))
-		default:
-			return fmt.Errorf("unknown log-format %q", context.GlobalString("log-format"))
-		}
-		return nil
-	}
-
-	// If the command returns an error, cli takes upon itself to print
-	// the error on cli.ErrWriter and exit.
-	// Use our own writer here to ensure the log gets sent to the right location.
-	cli.ErrWriter = &FatalWriter{cli.ErrWriter}
-	if err := app.Run(os.Args); err != nil {
-		fatal(err)
-	}
-}
-
-type FatalWriter struct {
-	cliErrWriter io.Writer
-}
-
-func (f *FatalWriter) Write(p []byte) (n int, err error) {
-	logrus.Error(string(p))
-	return f.cliErrWriter.Write(p)
-}
